completing assignment #4: Practicing Property & Event Binding and View Encapsulation
adding GameControlComponent to app (my-first-app) containing two buttons that start and stop an interval that runs every second and emitting everySecond event with the number of seconds that the interval is running.

adding EvenComponent and OddComponent to app (my-first-app), both receiving a number array from the app component and appearing for every number in the array passing the number to the respective component using property binding, presenting the number in the screen.

inside the app component (my-first-app) - adding an odd and even number arrays together with a function that fires every time the everySecond event is emitting and pushing the number that comes with the event into one of the arrays with respect to their names.

adding previous assignments as comments to their respective apps to have easier access to them in the future.

arranging code in previous assignments for easier reading.