connecting between shopping item and shopping detail to show the details of the selected recipe item, and adding the functionality to the add button, inside the shopping edit component
adding Form Module to app module imports.

adding a click event to recipe item component that in turn emitting onSelectRecipe event that is calling sendRecipe function with the recipeElement as an argument.
in turn, sendRecipe function is emitting onRecipeSent event with the recipe chosen as an argument which is assigned to the selectedRecipe variable inside recipes component.

adding *ngIf to app-recipe-detail based on weather selectedRecipe is assigned a value or not, and if not (else) it renders the infoText ng-template with some text.
if selectedRecipe is assigned a value, it is passed to  app-recipe-detail as recipe and it's values are being presented instead of the hard coded text.

adding a references to both name input and amount input and adding a click event to the add button inside shopping-edit.component template, that calls addIngredient function which, in turn, emitting the onIngredientAddition event with the values from the input references as a new Ingredient type which is passed to addingToIngredients function inside shopping-list.component that adds the new ingredient to the ingredients array.